Move evidence view model to correct namespace
